+++ packages/jac/src/middlewares/index.ts
@@ -1,4 +1,4 @@
-export { flattenToAxiosJsonApiQuery } from './flatten-to-axios-json-api-query'
+export { flatJsonApiQuery } from './flat-json-api-query'
 export { parseJsonApiError } from './parse-json-api-error'
 export { parseJsonApiResponse } from './parse-json-api-response'
 export { setJsonApiHeaders } from './set-json-api-headers'