+++ packages/jac/src/middlewares/parse-json-api-error.ts
@@ -1,6 +1,5 @@
-import { AxiosError } from 'axios'
+import { FetcherError, HTTP_STATUS_CODES } from '@distributedlab/fetcher'
 
-import { HTTP_STATUS_CODES } from '@/enums'
 import { errors } from '@/errors'
 import { JsonApiError } from '@/errors'
 import { JsonApiResponseErrors } from '@/types'
@@ -10,7 +9,7 @@ import { JsonApiResponseErrors } from '@/types'
  * Needed to handle on client side different behavior based on error type
  */
 export const parseJsonApiError = (
-  error: AxiosError<JsonApiResponseErrors>,
+  error: FetcherError<JsonApiResponseErrors>,
 ): JsonApiError => {
   const status = error?.response?.status