+++ packages/jac/src/middlewares/set-json-api-headers.test.ts
@@ -1,8 +1,8 @@
 import { setJsonApiHeaders } from './set-json-api-headers'
 
 describe('setJsonapiHeaders', () => {
-  it('should set proper set of headers', () => {
-    const headers = setJsonApiHeaders({ headers: {} })
+  test('should set proper set of headers', () => {
+    const headers = setJsonApiHeaders({})
 
     expect(headers).toStrictEqual({
       'Content-Type': 'application/vnd.api+json',